+++ plugins/general/index.js
@@ -6,6 +6,8 @@ import {plugins} from '../../modules/plugins';
 // Other
 import chalk from 'chalk';
 import packageJSON from '../../package';
+import request from 'request';
+import fs from 'fs';
 
 function aboutCommand(user, userID, channelID) {
     bot.sendMessage({
@@ -31,11 +33,56 @@ function commandsCommand(user, userID, channelID) {
 
             for (let command in plugin.commands) {
                 if (plugin.commands.hasOwnProperty(command)) {
+                    // Create a list with enabled synonyms for this command
+                    let synonyms = [];
+
+                    // Check plugins default synonyms
+                    if (plugin.commands[command].synonyms) {
+                        synonyms = plugin.commands[command].synonyms;
+                    }
+
+                    if (synonyms.indexOf(command) < 0) {
+                        synonyms.unshift(command);
+                    }
+
+                    // Check config.json for synonyms
+                    if (
+                        configModule.get().plugins
+                        && configModule.get().plugins[plugin.name]
+                        && configModule.get().plugins[plugin.name].commands
+                        && configModule.get().plugins[plugin.name].commands[command]
+                        && configModule.get().plugins[plugin.name].commands[command].synonyms
+                    ) {
+                        for (let synonym in configModule.get().plugins[plugin.name].commands[command].synonyms) {
+                            if (configModule.get().plugins[plugin.name].commands[command].synonyms.hasOwnProperty(synonym)) {
+                                if (configModule.get().plugins[plugin.name].commands[command].synonyms[synonym].enabled) {
+                                    if (synonyms.indexOf(synonym) < 0) {
+                                        synonyms.push(synonym);
+                                    }
+                                } else if (configModule.get().plugins[plugin.name].commands[command].synonyms[synonym].enabled === false) {
+                                    const index = synonyms.indexOf(synonym);
+                                    if (index >= 0) {
+                                        synonyms.splice(index, 1);
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                    // Compile string
                     string += '`' + configModule.get().globalCommandPrefix
                         + (pluginCommandPrefix ? pluginCommandPrefix + ' ' : '')
-                        + command + '`'
-                        + ' ' + plugin.commands[command].description
-                        + '\n';
+                        + synonyms[0] + '`'
+                        + ' ' + plugin.commands[command].description;
+
+                    synonyms.shift();
+
+                    if (synonyms.length > 0) {
+                        string += ' (synonyms: `' + synonyms.join('`, `') + '`)'
+                            + '\n';
+                    } else {
+                        string += '\n';
+                    }
                 }
             }
 
@@ -49,64 +96,66 @@ function commandsCommand(user, userID, channelID) {
     });
 }
 
-function configCommand(user, userID, channelID, message) {
-    message = message.split(' ');
-
-    // Check if a property and a new value is present
-    if (message.length < 2) {
-        bot.sendMessage({
-            to: channelID,
-            message: 'Example use of this command: `!config credentials.name The Best Bot`',
-        });
-
-        return false;
-    }
-
-    let changingProperty = message[0].split('.');
-    message.shift(); // Remove the property from the message
-    let newValue = message.join(' ');
-    if (newValue === 'true') { // Make a boolean to a true boolean type
-        newValue = true;
-    } else if (newValue === 'false') { // Make a boolean to a true boolean type
-        newValue = false;
-    } else if (isNaN(newValue)) { // Make a number to a true number type
-        newValue = newValue;
-    } else {
-        newValue = Number(newValue);
-    }
-
-    // Create a object
-    let newConfig = {};
-    for (let i = changingProperty.length - 1; i >= 0; i--) {
-        if (i === changingProperty.length - 1) {
-            newConfig[changingProperty[i]] = newValue;
-        } else {
-            let lastSegment = JSON.parse(JSON.stringify(newConfig));
-            newConfig = {};
-            newConfig[changingProperty[i]] = lastSegment;
-        }
-    }
-
-    // Save the new config
-    configModule.save(newConfig, error => {
-        if (error) {
-            console.error(error);
-            bot.sendMessage({
-                to: channelID,
-                message: 'There was a problem with saving the new config.',
-            });
-
-            return false;
-        }
-
-        bot.sendMessage({
-            to: channelID,
-            message: 'Config successfully changed.',
-        });
-    });
-}
+// function configCommand(user, userID, channelID, message) {
+//     message = message.split(' ');
+//
+//     // Check if a property and a new value is present
+//     if (message.length < 2) {
+//         bot.sendMessage({
+//             to: channelID,
+//             message: 'Example use of this command: `!config credentials.name The Best Bot`',
+//         });
+//
+//         return false;
+//     }
+//
+//     let changingProperty = message[0].split('.');
+//     message.shift(); // Remove the property from the message
+//     let newValue = message.join(' ');
+//     if (newValue === 'true') { // Make a boolean to a true boolean type
+//         newValue = true;
+//     } else if (newValue === 'false') { // Make a boolean to a true boolean type
+//         newValue = false;
+//     } else if (isNaN(newValue)) { // Make a number to a true number type
+//         newValue = newValue;
+//     } else {
+//         newValue = Number(newValue);
+//     }
+//
+//     // Create a object
+//     let newConfig = {};
+//     for (let i = changingProperty.length - 1; i >= 0; i--) {
+//         if (i === changingProperty.length - 1) {
+//             newConfig[changingProperty[i]] = newValue;
+//         } else {
+//             let lastSegment = JSON.parse(JSON.stringify(newConfig));
+//             newConfig = {};
+//             newConfig[changingProperty[i]] = lastSegment;
+//         }
+//     }
+//
+//     // Save the new config
+//     configModule.save(newConfig, error => {
+//         if (error) {
+//             console.error(error);
+//             bot.sendMessage({
+//                 to: channelID,
+//                 message: 'There was a problem with saving the new config.',
+//             });
+//
+//             return false;
+//         }
+//
+//         bot.sendMessage({
+//             to: channelID,
+//             message: 'Config successfully changed.',
+//         });
+//     });
+// }
 
 function killCommand() {
+    bot.disconnect();
+
     console.log(chalk.yellow('The Discord Bot API got stopped through the kill command.'));
     console.log(''); // Empty line
     process.exit();
@@ -199,7 +248,10 @@ function opCommand(user, userID, channelID, message) {
         return false;
     }
 
-    configModule.op(message[0], message[1], error => {
+    const operatorID = message[0];
+    message.shift();
+
+    configModule.op(operatorID, message, error => {
         if (error) {
             bot.sendMessage({
                 to: channelID,
@@ -226,7 +278,10 @@ function deopCommand(user, userID, channelID, message) {
         return false;
     }
 
-    configModule.deop(message[0], message[1], error => {
+    const operatorID = message[0];
+    message.shift();
+
+    configModule.deop(operatorID, message, error => {
         if (error === 'no such permission') {
             bot.sendMessage({
                 to: channelID,
@@ -321,6 +376,68 @@ function reloadCommand(user, userID, channelID) {
     });
 }
 
+function setAvatar(base64, channelID) {
+    bot.editUserInfo({
+        avatar: base64,
+        password: configModule.get().credentials.password,
+    }, () => {
+        bot.sendMessage({
+            to: channelID,
+            message: 'Avatar successfully changed.',
+        });
+    });
+}
+
+function avatarCommand(user, userID, channelID, message) {
+    const path = message.split(' ')[0];
+    if (path.length < 1) {
+        bot.sendMessage({
+            to: channelID,
+            message: 'You have to add a relative path or an url to the new avatar.',
+        });
+        return false;
+    }
+
+    configModule.avatar(path, error => {
+        if (error) {
+            bot.sendMessage({
+                to: channelID,
+                message: 'There was an error with saving the new avatar to your `config.json`.',
+            });
+            return false;
+        }
+
+        // Set the avatar of the bot to the one defined in the configModule.json
+        if (configModule.get().credentials.avatar && configModule.get().credentials.avatar !== null) {
+            const reg = new RegExp(/^(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&\/\/=]*)$/, 'gi');
+            if (reg.test(configModule.get().credentials.avatar)) {
+                request({
+                    url: configModule.get().credentials.avatar,
+                    encoding: null,
+                }, (error, response, body) => {
+                    if (!error && response.statusCode == 200) {
+                        setAvatar(new Buffer(body).toString('base64'), channelID);
+                    } else {
+                        console.log(chalk.red('The avatar could not be set. Make sure the path is correct.'));
+                    }
+                });
+            } else {
+                setAvatar(fs.readFileSync(configModule.get().credentials.avatar, 'base64'), channelID);
+            }
+        } else if (configModule.get().credentials.avatar === null) {
+            bot.editUserInfo({
+                avatar: null,
+                password: configModule.get().credentials.password,
+            }, () => {
+                bot.sendMessage({
+                    to: channelID,
+                    message: 'Avatar successfully changed.',
+                });
+            });
+        }
+    });
+}
+
 let plugin = {
     name: 'general',
     commands: {
@@ -331,46 +448,72 @@ let plugin = {
         commands: {
             fn: commandsCommand,
             description: 'Shows all available commands',
+            synonyms: [
+                'help',
+            ],
         },
         kill: {
             fn: killCommand,
             description: 'Stops the bot',
+            requirePermission: true,
+            synonyms: [
+                'stop',
+            ],
         },
         userid: {
             fn: userIDCommand,
             description: 'Returns the ID of the user',
+            synonyms: [
+                'me',
+            ],
         },
         enable: {
             fn: enableCommand,
             description: 'Enables a plugin',
+            requirePermission: true,
         },
         // restart: {
         //     fn: restartCommand,
         //     description: 'Restarts the bot',
+        //     requirePermission: true,
         // },
         rename: {
             fn: renameCommand,
             description: 'Renames the bot',
+            requirePermission: true,
         },
         op: {
             fn: opCommand,
             description: 'Adds a permission to a user',
+            requirePermission: true,
         },
         deop: {
             fn: deopCommand,
             description: 'Removes a permission from a user',
+            requirePermission: true,
         },
         prefix: {
             fn: prefixCommand,
             description: 'Changes to global command prefix',
+            requirePermission: true,
         },
         owner: {
             fn: ownerCommand,
             description: 'Changes the owner of the bot',
+            requirePermission: true,
         },
         reload: {
             fn: reloadCommand,
-            description: 'Reloads the `config.json`',
+            description: 'Reloads your `config.json`',
+            requirePermission: true,
+            synonyms: [
+                'refresh',
+            ],
+        },
+        avatar: {
+            fn: avatarCommand,
+            description: 'Give the bot an avatar',
+            requirePermission: true,
         },
     },
 };